# Optimize zend index after indexing and aggregate fields