--- a/Resources/config/adapter_zendlucene.xml
+++ b/Resources/config/adapter_zendlucene.xml
@@ -14,6 +14,7 @@
             <argument>%massive_search.adapter.zend_lucene.basepath%</argument>
             <argument>%massive_search.adapter.zend_lucene.hide_index_exception%</argument>
             <argument>%massive_search.adapter.zend_lucene.encoding%</argument>
+            <tag name="kernel.event_listener" event="massive_search.index_rebuild" method="optizeIndexAfterRebuild" priority="-999" />
         </service>
 
     </services>