--- a/composer.json
+++ b/composer.json
@@ -19,6 +19,7 @@
         "elasticsearch/elasticsearch": "~1.3",
         "symfony-cmf/testing": "dev-master",
         "symfony/filesystem": "~2.4",
+        "symfony/expression-language": "~2.4",
         "symfony/console": "~2.5",
         "symfony/expression-language": "~2.4",
         "phpspec/prophecy-phpunit": "~1.0",