This PR improves the speed of the zend lucene search:

- Optimizes the index after an index rebuild
- Aggregates the indexed field values into a single (unstored) field

With a test this reduces the query time from 1.1s to 0.2s
Optimizing the index saves ~300ms
Aggregating the fields saves ~600ms

Original search:

````
$ ./app/console massive:search:query "anton" 
168 result(s) in 3.827782
````

With optimization and aggregation:

````
$ ./app/console massive:search:query "anton" 
168 result(s) in 1.634348s
````